Run Async code (side effect) Inside  the Component
As we do not have a backend server using node.js or PHP, we need to transform the data in the redux file. If we had the backend server, this transformation should be performed on the backend. In this commit, we started adding items to the cart from inside the ProductItem component.js without using the action creator.

I worked on the add-to-cart handler in ProductItem.js. First, I used the useSelector from react-redux to get the cart state from the store (cartSlice: items and totalQuantity). Then, in the addToCartHandler, I created a new constant, newTotalQuantity, without mutating the state because we cannot change it outside the reducer methods. Similarly, I created an updated items constant, updatedItems, by copying the items in the cart with the slice method that creates a new array with the existing objects. I then copied the existing item into a new object, updatedItem, and updated its properties without manipulating the Redux store. If I were to update an existing item without creating a copy first, I would be manipulating the Redux store. I then found the index of the existing item and replaced it with the updated item. If the item did not exist, I pushed a new object to my updated items array. Finally, I created a new object for the new cart with the updated total quantity and items and dispatched a new action using the replaceCart reducer in the cart-slice.js. The payload was an object with a total quantity and items key (const newCart in ProductItem.js). This action was dispatched in ProductItem.js. I am not sending an HTTP request yet. In summary, all this code just transforms data in async code and is side-effect free. In such a case, we should use reducers. If we use async code or code with side-effects, we should prefer action creators or components and never use reducers.